+++ modules/matrix-bridge/test/matrix-event-handler-test.js
@@ -2,6 +2,7 @@
 
 const assert = require('assert');
 const fixtureLoader = require('gitter-web-test-utils/lib/test-fixtures');
+const mongoUtils = require('gitter-web-persistence-utils/lib/mongo-utils');
 const MatrixEventHandler = require('../lib/matrix-event-handler');
 const store = require('../lib/store');
 const chatService = require('gitter-web-chats');
@@ -107,7 +108,20 @@ describe('matrix-event-handler', () => {
     describe('handleChatMessageCreateEvent', () => {
       const fixture = fixtureLoader.setupEach({
         userBridge1: {},
-        troupe1: {}
+        user1: {},
+        troupe1: {},
+        troupeWithThreads1: {},
+        messageParent1: {
+          user: 'user1',
+          troupe: 'troupeWithThreads1',
+          text: 'some parent message'
+        },
+        messageThread1: {
+          parent: 'messageParent1',
+          user: 'user1',
+          troupe: 'troupeWithThreads1',
+          text: 'some parent message'
+        }
       });
 
       beforeEach(() => {
@@ -128,6 +142,101 @@ describe('matrix-event-handler', () => {
         const messages = await chatService.findChatMessagesForTroupe(fixture.troupe1.id);
         assert.strictEqual(messages.length, 1);
         assert.strictEqual(messages[0].text, 'my matrix message');
+        assert.strictEqual(messages[0].virtualUser.externalId, 'alice:localhost');
+        assert.strictEqual(messages[0].virtualUser.displayName, 'Alice');
+      });
+
+      it('reply to message starts threaded conversation', async () => {
+        const matrixMessageEventId = `$${fixtureLoader.generateGithubId()}`;
+        const eventData = createEventData({
+          type: 'm.room.message',
+          content: {
+            body: 'my matrix reply',
+            'm.relates_to': {
+              'm.in_reply_to': {
+                event_id: matrixMessageEventId
+              }
+            }
+          }
+        });
+        await store.storeBridgedRoom(fixture.troupeWithThreads1.id, eventData.room_id);
+        // Replying to a parent message
+        await store.storeBridgedMessage(
+          fixture.messageParent1.id,
+          eventData.room_id,
+          matrixMessageEventId
+        );
+
+        await matrixEventHandler.onEventData(eventData);
+
+        const messages = await chatService.findChatMessagesForTroupe(
+          fixture.troupeWithThreads1.id,
+          { includeThreads: true }
+        );
+        assert.strictEqual(messages.length, 3);
+        assert.strictEqual(messages[2].text, 'my matrix reply');
+        assert(mongoUtils.isLikeObjectId(messages[2].parentId, fixture.messageParent1.id));
+      });
+
+      it("reply to message in thread that isn't the parent puts it in the thread correctly", async () => {
+        const matrixMessageEventId = `$${fixtureLoader.generateGithubId()}`;
+        const eventData = createEventData({
+          type: 'm.room.message',
+          content: {
+            body: 'my matrix reply',
+            'm.relates_to': {
+              'm.in_reply_to': {
+                event_id: matrixMessageEventId
+              }
+            }
+          }
+        });
+        await store.storeBridgedRoom(fixture.troupeWithThreads1.id, eventData.room_id);
+        // Replying to a message in the thread
+        await store.storeBridgedMessage(
+          fixture.messageThread1.id,
+          eventData.room_id,
+          matrixMessageEventId
+        );
+
+        await matrixEventHandler.onEventData(eventData);
+
+        const messages = await chatService.findChatMessagesForTroupe(
+          fixture.troupeWithThreads1.id,
+          { includeThreads: true }
+        );
+        assert.strictEqual(messages.length, 3);
+        assert.strictEqual(messages[2].text, 'my matrix reply');
+        assert(mongoUtils.isLikeObjectId(messages[2].parentId, fixture.messageParent1.id));
+      });
+
+      it('When we receive Matrix image upload, creates Gitter message linking the image in Gitter room', async () => {
+        const eventData = createEventData({
+          type: 'm.room.message',
+          content: {
+            body: 'my-image.jpeg',
+            info: {
+              h: 807,
+              mimetype: 'image/jpeg',
+              size: 176995,
+              thumbnail_info: [Object],
+              thumbnail_url: 'mxc://my.matrix.host/DSVnwDtcZeNoBEjUnsFxdelN',
+              w: 1217
+            },
+            msgtype: 'm.image',
+            url: 'mxc://my.matrix.host/yjyeYJIBdJGkYvYoLuvPfBuS'
+          }
+        });
+        await store.storeBridgedRoom(fixture.troupe1.id, eventData.room_id);
+
+        await matrixEventHandler.onEventData(eventData);
+
+        const messages = await chatService.findChatMessagesForTroupe(fixture.troupe1.id);
+        assert.strictEqual(messages.length, 1);
+        assert.strictEqual(
+          messages[0].text,
+          '[my-image.jpeg](http://localhost:18008/_matrix/media/v1/download/my.matrix.host/yjyeYJIBdJGkYvYoLuvPfBuS)\n[![my-image.jpeg](http://localhost:18008/_matrix/media/v1/download/my.matrix.host/DSVnwDtcZeNoBEjUnsFxdelN)](http://localhost:18008/_matrix/media/v1/download/my.matrix.host/yjyeYJIBdJGkYvYoLuvPfBuS)'
+        );
         assert.strictEqual(messages[0].virtualUser.externalId, 'alice:localhost');
         assert.strictEqual(messages[0].virtualUser.displayName, 'Alice');
       });